Copy "docker cp" command from docker/cli project
Part of the workflow of the terraform runner involves copying/writing a terraform file to the container's volume. I was originally intending to do this with the Docker GO SDK's CopyToContainer() function. However, I ran into several issues related to "unexpected EOF" of the tar file. After spending numerous hours trying to debug the issue I realized that the "docker cp" command worked perfectly for my purposes. So I decided to reuse the logic from the "docker cp" command, and it worked!

The crazy thing is that the docker cp command itself uses the Docker GO SDK's CopyToContainer function (obviously!), but it seems like they do something extra to make sure they don't have an "unexpected EOF".

For now I have reused the code wholesale, but the goal is to slowly strip the "docker cp" code until I get to the core of what I really need.